requirements.txt
absl-py==2.1.0
albucore==0.0.19
albumentations==1.4.20
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.1
blis==1.0.1
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
crypto==1.4.1
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.7
decorator==5.1.1
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.4
filterpy==1.4.5
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
frozendict==2.4.6
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==23.0.0
h5py==3.11.0
html5lib==1.1
huggingface-hub==0.26.2
idna==3.6
imageio==2.34.1
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.10
Naked==0.1.32
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.59.1
numpy==2.0.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
peewee==3.17.7
pillow==10.3.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pymongo==4.10.1
PyMuPDF==1.24.13
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==308
PyYAML==6.0.1
pyzmq==26.2.0
regex==2024.9.11
requests==2.31.0
rich==13.7.1
safetensors==0.4.5
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
stringzilla==3.10.6
sympy==1.13.1
tbb==2021.12.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminaltables==3.1.10
textblob==0.18.0.post0
tf_keras==2.18.0
thinc==8.3.2
threadpoolctl==3.4.0
tifffile==2024.4.24
tokenizers==0.20.1
torch==2.5.1
torchdata==0.6.0
torchtext==0.15.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.46.1
typer==0.12.5
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
utils==1.0.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yfinance==0.2.48
